{noup} scsi: hisi_sas: Add v2 hw debugfs support
This is just a hack to enable debugfs for v2 hw, to enabled
basic debugfs feature when hi1620 is not available for testing.

This will break boot on v1 hw if debugfs is enabled on that same
platform, as we would get a NULL-pointer deference.

Actually added v2 hw support should not be too difficult in reality,
considering the hw-agnostic nature of the debugfs support we added.

Signed-off-by: John Garry <john.garry@huawei.com>