scsi: libsas: Use proper method to calculate linkrates in sas_ex_get_linkrate()
Currently we calculate the device linkrate in sas_ex_get_linkrate()
using the @parent_phy argument.

The significance of this parent PHY is that it is just the first PHY
which we detect the device on.

As such, the code implies that that the negotiated linkrate of every
constituent PHY in an expander's wideport is the same. This is not
true.

This patch updates sas_ex_get_linkrate() to use a proper method to get
the linkrates.

Signed-off-by: John Garry <john.garry@huawei.com>