scsi: libsas: Support SATA PHY connection rate unmatch fixing for root port revalidation
In commit cec9771d2e95 ("scsi: libsas: Support SATA PHY connection rate
unmatch fixing during discovery"), we added the ability to check that the
maximum pathway from the initiator to a SATA target PHY does not exceed
that PHY linkrate; if such a condition did occur, we would attempt to lower
the SATA PHY linkrate.

This only came in affect for when the SATA device was initially being
discovered, and not thereafter.

However the maximum pathway condition can fail after the SATA device is
initially discovered. Lowering the linkrate of any root PHY below the SATA
PHY linkrate would be such a condition.

This change ensures that we maintain the maximum pathway for any SATA
PHY connected to a root port expander for when the root port expander
minimum linkrate is lowered.

It works by updating each domain device's linkrates after we find a
change in the root port linkrates. Then we check the maximum pathway
condition for each SATA device - if it is not satisfied, then the device
is deregistered and SATA PHY is reset at the appropiated rates.

When we deregister the PHY, we lose the device.

The reason we must deregister the PHY and lose the device is because
some HBAs must record the maximum connection rate of a device; if this
is lowered, then the HBA LLDD must be informed. The only safe was to do
this is to lose and then rediscover the device.

The scenario of cascaded expanders is not supported.

Signed-off-by: John Garry <john.garry@huawei.com>