scsi: hisi_sas: Pass rates straight through in hisi_sas_phy_set_linkrate()
Now that libsas sets the rates for both minimum and maximum, the LLDD
no longer needs to do it.

Other drivers have been audited and do not need to make the same change.

Also increase the sleep between phy down and phy up, so that libsas
has time to process the phy down event fully.

Signed-off-by: John Garry <john.garry@huawei.com>