scsi: libsas: Add root port min and max linkrates
Currently we only record the linkrate of a root port, which is calculated
as the negotiated linkrate of the first member PHY.

This is not strictly the correct approach, as a port may be formed from
members with different negotiated linkrates.

This change adds support for recording the min and max linkrate of a
root port.

Some other variables are renamed/introduced for code consistency.

Signed-off-by: John Garry <john.garry@huawei.com>