scsi: libsas: Support SATA PHY connection rate unmatch fixing for SMP PHY
Currently we support SATA PHY connection rate unmatch fixing for when
the root port device minimum linkrate changes.

In this patch, we support matching for when the rate of an SMP PHY changes
(according to SAS terminology, an "SMP" PHY is an expander PHY attached to
another expander PHY).

The same method is used as for root expander: that is, in the
revalidation phase update all device linkrates from the bcast expander,
check the min pathway condition per device, and fixup as necessary.

Signed-off-by: John Garry <john.garry@huawei.com>