scsi: libsas: Process all PHY change count events
In processing a broadcast event, when a change is detected in an expander
phy change count, we rediscover the phy and then update the phy change
count in software with the value read from the SMP phy discovery helper.
In doing so, all phy change events are essentially cleared.

This may have worked for before when libsas events were not generated
dynamically. Back then, libsas would clear a revalidation (broadcast)
event flag, clear expander and phy change count events, and then conclude
all revalidation is complete.

However, with the change to process events dynamically, one by one, we
still clear the expander and phy change counts for each revalidation.
This causes problems when an event is generated during a revalidation, in
that the current revalidation may clear all expander or phy change counts,
and the event may be missed.

Remedy this by incrementing any phy's change count only once per
revalidation.

To solve the tricky situation of updating expander change count,
keep an expander's change count set at -1 until all phy changes are
processed.

We inject broadcast events to clear the phy change events. It should
only be necessary to do this for root port phys (since they do not
generate broadcast events), but there is not much cost to always
do it and also help avoid obscure corner cases.

Signed-off-by: John Garry <john.garry@huawei.com>