{topost} scsi: hisi_sas: Use reply map for v2 hw
With the aim of boosting performance, switch v2 hw to using reply map, i.e.
pseduo blk-mq.

This really helps the environment with we're running a distro with
irqbalance running. The reason being that the irqbalance can fiddle with
the completion queue affinity, such that we random cpus servicing the
CQs. It's not much better without irqbalance, where it is always cpu0.

Distro read fio performance for 6x SAS SSDs goes from ~450K IOPS ->
650K IOPS.

To continuing support manually setting the irq affinity, introduce module
parameter auto_affine_interrupts, which is default on.

Signed-off-by: John Garry <john.garry@huawei.com>