iommu/arm-smmu-v3: Don't update cons for when polling MSI is supported
For when polling MSI is supported, we don't really need to update cmdq
cons pointer. We only maintain it so that we know that the CMD_SYNC is
consumed in non-MSI polling mode.

Indeed, maintaining cons and also the 64b read of cmdq->q.llq.val can eat
an appreciable amount of cycles, so only did it when necessary (non-MSI
polling mode).

Signed-off-by: John Garry <john.garry@huawei.com>