{topost} scsi: hisi_sas: avoid accessing to ssp task for SMP IOs
In function hisi_sas_slot_task_free(), it accesses to ssp task for
non-ata task. But if it is smp task here, it accesses to wrong structure
though it may not cause any issue.
To avoid it, only access to ssp task when slot->n_elem_dif is not 0
which indicates this is ssp task.

Signed-off-by: Xiang Chen <chenxiang66@hisilicon.com>
Signed-off-by: John Garry <john.garry@huawei.com>