{topost} scsi: hisi_sas: Add BIST support for phy FFE
Add BIST support for phy FFE(Feed forward equalizer) setting.

Through the new debugfs interface, the user can configure FFE if they want.

FFE are a kind of parameter which used for link layer of board. They will
affect the link quality between the SAS controller and the backplane.In the
BIST test, FFE parameter interface is provided to assist board tester in
optimizing link parameters.

The modification of the FFE parameter will affect the test after BIST or
the normal running of the board, so user should save FFE's initial value
before modify it, and restore them after BIST test if necessary.

Signed-off-by: Luo Jiaxing <luojiaxing@huawei.com>
Signed-off-by: John Garry <john.garry@huawei.com>