{topost} scsi: hisi_sas: Do not modify upper fields of PROG_PHY_LINK_RATE reg
When we update register of PROG_PHY_LINK_RATE to set linkrate for phy, we
used a hard-coded initial value instead of getting the current value from
register. We had assumed that this register would not be modified, but in
fact it was partially modified in new version of hardware. So hard-coded
value we used change default value of register to a wrong setting and make
SAS controller can not change linkrate for phy at new version of hardware.

So we delete hard-coded value and always read the latest value of register
before we update part of it.

Signed-off-by: Luo Jiaxing <luojiaxing@huawei.com>
Signed-off-by: John Garry <john.garry@huawei.com>