blk-mq: Lockout tagset iterator when exiting elevator
All queues associated with a tagset are frozen when exiting an elevator.
This is to ensure that one queue running blk_mq_queue_tag_busy_iter()
cannot hold a stale request reference for the queue who is exiting the
elevator.

However, there is nothing to stop blk_mq_all_tag_iter() being run for
the tagset, and, again, getting hold of stale request reference. A kasan
UAF can be triggered for this scenario:

To avoid this, reject the tagset iterators when the queue is exiting
the elevator.

This should not break any semantics in blk_mq_all_tag_iter(), as, since
all queues are frozen, there should be no tags to iterate.

Signed-off-by: John Garry <john.garry@huawei.com>