{topost} scsi: hisi_sas: directly snapshot registers when FLR
Debugfs dump should be executed before FLR run for we have to dump some
registers before reset by FLR. So it's wrong to queue debugfs dump
work when running FLR work because these two work queue is in the same
workqueue. It mean that Debugfs dump work is always execute after FLR and
get data which is reset.

Therefore, directly snapshot registers in the FLR work instead of queue a
work.

Signed-off-by: Jianqin Xie <xiejianqin@hisilicon.com>
Signed-off-by: Luo Jiaxing <luojiaxing@huawei.com>
Signed-off-by: John Garry <john.garry@huawei.com>