block: Rename BLKDEV_MAX_RQ -> BLKDEV_DEFAULT_RQ
It is a bit confusing that there is BLKDEV_MAX_RQ and MAX_SCHED_RQ, as
the name BLKDEV_MAX_RQ would imply the max requests always, which it does
not.

Rename to BLKDEV_DEFAULT_RQ, matching it's usage - the default number of
requests assigned when allocating a request queue.

Signed-off-by: John Garry <john.garry@huawei.com>