null_blk: Add pseudo-managed interrupts support
blk-mq supports draining HW queues contexts of active requests for when the
underlying HW uses managed interrupts signalling and all associated CPUs
are offlined.

Testing this feature may be difficult as appropiate HW may not be available
or certain configurations may also be difficult to create.

Add pseudo-managed interrupt support by making requests timeout when no
CPUs are online in the hctx cpumask at completion.

A new irqmode is added, NULL_IRQ_MANAGED. This mode is based on a timer, as
a time window needs to be created between "submission and completion" for
the CPU(s) to go offline, which other modes like softirq would not provide.

Signed-off-by: John Garry <john.garry@huawei.com>