scsi: hisi_sas: replace del_timer() with del_timer_sync()
For the timer of sas_task->slow_task, it may be blocked in function
hisi_sas_task_exec() as there are some mutex/spinlock, so it is possible
that the timer will be expired and callback running when calling
del_timer(). It is also the same for slot->internal_abort_timer. For the
timer of hisi_hba, it also may have the same issue as the link timeout
may occur anytime. Though the probability is very slow, replace
del_timer() with del_timer_sync() to avoid it except some special
situations (del_timer() is required in interrupt context).
Also timer_pending() is not necessary as it is already checked in function
del_timer_sync() and del_timer().

Signed-off-by: Xiang Chen <chenxiang66@hisilicon.com>
Signed-off-by: John Garry <john.garry@huawei.com>