scsi: hisi_sas: increase debugfs_dump_index after dump is completed
The debugfs_dump_index increase before DFX dump data. This is incorrect
because dump array is obtained based on debugfs_dump_index during the dump.
As a result, dump array of index 0 is never used. When the last entry of
dump array is obtained, memory overwriting occurs.

Therefore, the operation of increasing debugfs_dump_index by 1 is moved
after the dump is completed.

Signed-off-by: Luo Jiaxing <luojiaxing@huawei.com>
Signed-off-by: John Garry <john.garry@huawei.com>