blk-mq-debugfs: Show active requests per queue for shared tags
Currently we always show the hctx.active member for per-hctx "active" file.

However this is not maintained for shared tags, and we instead keep a
record of the number active requests per request queue - see commit
f1b49fdc1c64 ("blk-mq: Record active_queues_shared_sbitmap per tag_set for
when using shared sbitmap).

Change for the the case of shared tags to note that the active requests are
per queue and use helper __blk_mq_active_requests() to retrieve the value.

Signed-off-by: John Garry <john.garry@huawei.com>