scsi: libsas: Add sas_execute_tmf()
Drivers using libsas have to issue their own TMFs, and the code to do this
is duplicated between drivers.

Add a common function to handle TMFs. This will be also used for sending
ATA commands, but use name "tmf" as the prupose is similar between SCSI
and ATA in the usage.

Signed-off-by: John Garry <john.garry@huawei.com>